// src/modules/invoice/facade/generate.invoice.facade.interface.ts
export interface GenerateInvoiceFacadeInputDto {
  name: string;
  document: string;
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  items: {
    id: string;
    name: string;
    price: number;
  }[];
}

export interface GenerateInvoiceFacadeOutputDto {
  id: string;
  name: string;
  document: string;
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  items: {
    id: string;
    name: string;
    price: number;
  }[];
  total: number;
}

export default interface GenerateInvoiceFacadeInterface {
  process(input: GenerateInvoiceFacadeInputDto ): Promise<GenerateInvoiceFacadeOutputDto>;
}